+++ CHANGELOG.md
@@ -6,9 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## Added
+- Support of dict (keys and values) copy in generic deepcopy
+
 ### Changed
 - Export csv to fit vectored_objects removal
 - All set variables in _display_angular config instead of just choice_variables
+- Strategy for scale and custom value scaling
 
 ## [0.3.5]
 ### Changed